pyproject.toml
[project]
name = "gtfs-validator-api"
version = "0.1.0"
description = "FastAPI wrapper for MobilityData GTFS Validator"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "uv",
    "aiohttp",
    "fastapi-mail",
    "markdown",
    "slowapi",
    "bcrypt",
    "jinja2",
    "python-dotenv",
    "google-cloud-firestore"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "types-requests",
    "starlette-openapi",
    "httpx",
]


